Create FoodOrder.py
This is a Python program that simulates a food order at a burger place (specifically The Habit Burger Grill). The user chooses a burger, and has the option to choose a beverage and fries. The capitalization of the inputs doesn't matter, as the .lower() is incorporated to make the application not case sensitive. However, the program does make sure that it does not break, as the program error checks each variable, and asks the question again if the user types the wrong input again. If the user does this 10 times for one variable, the program closes.